updating.. for new EMR&al2023

+++ packer-files/scripts/samtools.sh
@@ -8,11 +8,7 @@ set -xe
 REPOSITORY_URL="https://github.com/samtools/samtools.git"
 HTSLIB_DIR="/opt/htslib"
 
-yum -y install \
-    autoconf \
-    gcc72 \
-    git \
-    ncurses-devel
+dnf -y install ncurses-devel
 
 if [ -z "$SAMTOOLS_VERSION" ]; then
     SAMTOOLS_VERSION="master";
@@ -21,14 +17,15 @@ fi
 
 mkdir -p /opt
 cd /opt
-git clone "$REPOSITORY_URL"
-cd samtools
-git checkout "$SAMTOOLS_VERSION"
-#git checkout "1.20"
-autoheader
-autoconf -Wno-syntax
-./configure --with-htslib="$HTSLIB_DIR"
-make -j "$(grep -c ^processor /proc/cpuinfo)"
-make install
+if [ ! -d samtools ]; then
+   git clone "$REPOSITORY_URL"
+   cd samtools
+   git checkout "$SAMTOOLS_VERSION"
+   autoheader
+   autoconf -Wno-syntax
+   ./configure --with-htslib="$HTSLIB_DIR"
+   make -j "$(grep -c ^processor /proc/cpuinfo)"
+   make install
+   rm -rf /opt/samtools
+fi
 
-rm -rf /opt/samtools
